LockOnPluginKK - Remove pointless code
